+++ include/esp_secure_cert_tlv_config.h
@@ -27,7 +27,7 @@ typedef enum esp_secure_cert_tlv_type {
     ESP_SECURE_CERT_TLV_SEC_CFG,
     ESP_SECURE_CERT_TEE_SEC_STORAGE_KEY_TLV,
     ESP_SECURE_CERT_SIGNATURE_BLOCK_TLV,
-    ESP_SECURE_CERT_TLV_TYPE_INTEGRITY,
+    ESP_SECURE_CERT_INTEGRITY_TLV,
     // Any new tlv types should be added above this
     ESP_SECURE_CERT_TLV_END = 50,
     //Custom data types